refactor: move some state code around
