+++ src/app/layout_manager.rs
@@ -872,11 +872,11 @@ pub struct BottomWidget {
     #[builder(default = None)]
     pub parent_reflector: Option<(WidgetDirection, u64)>,
 
-    /// Top left corner when drawn, for mouse click detection.  (x, y)
+    /// Top left corner when drawn, for mouse click detection. (x, y)
     #[builder(default = None)]
     pub top_left_corner: Option<(u16, u16)>,
 
-    /// Bottom right corner when drawn, for mouse click detection.  (x, y)
+    /// Bottom right corner when drawn, for mouse click detection. (x, y)
     #[builder(default = None)]
     pub bottom_right_corner: Option<(u16, u16)>,
 }
