+++ src/app/widgets.rs
@@ -11,3 +11,12 @@ pub use disk_table::*;
 
 pub mod cpu_graph;
 pub use cpu_graph::*;
+
+pub mod net_graph;
+pub use net_graph::*;
+
+pub mod mem_graph;
+pub use mem_graph::*;
+
+pub mod battery_widget;
+pub use battery_widget::*;
